+++ src/main/java/com/aaax/server/config/redis/RedisConfig.java
@@ -1,19 +1,17 @@
 package com.aaax.server.config.redis;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import com.aaax.core.utils.RedisUtil;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.core.RedisOperations;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.serializer.GenericJackson2JsonRedisSerializer;
-import org.springframework.data.redis.serializer.GenericToStringSerializer;
+import org.springframework.data.redis.serializer.GenericJacksonJsonRedisSerializer;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
+import tools.jackson.databind.DeserializationFeature;
+import tools.jackson.databind.json.JsonMapper;
 
 
 @Configuration
@@ -40,20 +38,12 @@ public class RedisConfig {
 
     @Bean
     public RedisTemplate<String, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
-        ObjectMapper objectMapper = new ObjectMapper()
-                .registerModule(new JavaTimeModule()) // Register the JavaTimeModule for Java 8 date/time types
-                ;
-        // 1.create redisTemplate
         RedisTemplate<String, Object> redisTemplate = new RedisTemplate<>();
-        // 2.set redisConnectionFactory
         redisTemplate.setConnectionFactory(redisConnectionFactory);
-        // 3.set serializer
-        GenericToStringSerializer genericToStringSerializer = new GenericToStringSerializer(Object.class);
-        // 6.GenericToStringSerializer，mapping
-        // 7.set value and object serializer
-        redisTemplate.setValueSerializer(genericToStringSerializer);
         redisTemplate.setKeySerializer(new StringRedisSerializer());
-        redisTemplate.setValueSerializer(new GenericJackson2JsonRedisSerializer(objectMapper));
+        redisTemplate.setValueSerializer(new GenericJacksonJsonRedisSerializer(JsonMapper.builder()
+                .disable(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES)
+                .build()));
         redisTemplate.afterPropertiesSet();
         return redisTemplate;
     }